refactor(tests): Organize tests module files
