+++ tests/e2e/mod.rs
@@ -1,0 +1,7 @@
+
+
+#[tokio::test]
+async fn test_single_request_get() -> anyhow::Result<()> {
+    assert!(false);
+    Ok(())
+}
