--- tests/mock_api/docker-compose.yml
@@ -1,12 +1,0 @@
-version: '3.8'
-services:
-  mockapi:
-    container_name: mockapi
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-      - "7777:7777"
-    
-
-
