+++ tests/mod.rs
@@ -1,1 +1,6 @@
-pub mod provider;
+pub mod integration;
+pub mod services;
+pub mod utils;
+
+#[cfg(feature = "e2e")]
+pub mod e2e;
