--- tests/provider/mod.rs
@@ -1,3 +1,0 @@
-pub mod factory_provider;
-pub mod integration;
-pub mod services;
