+++ tests/setup/docker-compose.yml
@@ -1,0 +1,18 @@
+version: '3.8'
+services:
+  mockapi:
+    build:
+      context: mockoon
+      dockerfile: Containerfile
+    container_name: mockoon
+    ports:
+      - "7777:7777"
+  
+  httpbin:
+    image: kennethreitz/httpbin
+    container_name: httpbin
+    ports:
+      - "8888:80"
+    
+
+
